# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RECALL API
  version: 0.1.0
  description: 个性化游戏生成与分享平台 MVP API。
servers:
  - url: http://127.0.0.1:8080
    description: Local development
paths:
...
  /api/v1/analytics/events:
    post:
      operationId: recordCreatorBehaviorEvent
      summary: 上报当前制作方的白名单前端行为事件
      description: 仅 app/all Surface 挂载；第一版只接受 creator.page_viewed，制作方和用户会话关联均由服务端认证上下文推导。
      tags: [Analytics]
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/Origin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatorAnalyticsEventRequest"
      responses:
        "200":
          description: 合法幂等重试，返回原事件
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsEventWriteResponse"
        "201":
          description: 新事件已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsEventWriteResponse"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
  /api/v1/admin/behavior-events:
    get:
      operationId: listAdminBehaviorEvents
      summary: 查询脱敏的用户行为事件
      description: 仅 app/all Surface 挂载，按 createdAt 和 id 倒序进行不透明游标分页。
      tags: [Analytics]
      parameters:
        - name: eventName
          in: query
          schema:
            $ref: "#/components/schemas/BehaviorEventName"
        - name: creatorId
          in: query
          schema:
            $ref: "#/components/schemas/ULID"
        - name: loginId
          in: query
          description: 仅通过 users.login_id 参数化精确联查，不搜索事件属性。
          schema:
            type: string
            minLength: 3
            maxLength: 32
        - name: gameId
          in: query
          schema:
            $ref: "#/components/schemas/ULID"
        - name: source
          in: query
          schema:
            type: string
            enum: [frontend, api, worker]
        - name: from
          in: query
          description: createdAt 的包含性下界。
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: createdAt 的排除性上界。
          schema:
            type: string
            format: date-time
        - name: cursor
          in: query
          description: 服务端返回的无填充 Base64URL v1 游标。
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: 行为事件列表；空结果返回 items 为空数组、nextCursor 为 null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBehaviorEventListResponse"
        "401":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
...
components:
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
    Origin:
      name: Origin
      in: header
      required: true
      schema:
        type: string
        format: uri
      example: http://127.0.0.1:5173
...
  responses:
    APIError:
      description: 请求失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
        requestId:
          type: string
    ULID:
      type: string
      pattern: "^[0-7][0-9A-HJKMNP-TV-Z]{25}$"
      minLength: 26
      maxLength: 26
    BehaviorEventName:
      type: string
      enum:
        - creator.page_viewed
        - creator.registered
        - creator.logged_in
        - game.created
        - game.version_created
        - asset.uploaded
        - generation.submitted
        - generation.succeeded
        - generation.failed
        - share.created
        - share.opened
        - play.started
        - play.completed
        - play.replayed
    CreatorAnalyticsEventRequest:
      type: object
      additionalProperties: false
      required: [eventName, clientEventId, properties]
      properties:
        eventName:
          type: string
          const: creator.page_viewed
        clientEventId:
          type: string
          format: uuid
          pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
        occurredAt:
          type: string
          format: date-time
        properties:
          $ref: "#/components/schemas/CreatorPageViewedProperties"
...
    CreatorPageViewedProperties:
      type: object
      additionalProperties: false
      required: [page]
      properties:
        page:
          type: string
          enum: [create, games, game-edit, game-preview, game-share, generation-progress, settings]
    PublicPlayProperties:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode:
          type: string
          const: public
    AnalyticsEventWriteResponse:
      type: object
      additionalProperties: false
      required: [data, requestId]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [eventId, duplicate]
          properties:
            eventId:
              $ref: "#/components/schemas/ULID"
            duplicate:
              type: boolean
        requestId:
          $ref: "#/components/schemas/ULID"
    AdminBehaviorEventListResponse:
      type: object
      additionalProperties: false
      required: [data, requestId]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [items, nextCursor]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AdminBehaviorEvent"
            nextCursor:
              type: [string, "null"]
              description: 无填充 Base64URL v1 游标；无更多结果时为 null。
        requestId:
          $ref: "#/components/schemas/ULID"
    AdminBehaviorEvent:
      type: object
      additionalProperties: false
      required:
        - id
        - eventName
        - source
        - actorType
        - creatorId
        - loginId
        - userSessionId
        - gameId
        - gameVersionId
        - generationRunId
        - shareId
        - playSessionId
        - requestId
        - properties
        - occurredAt
        - createdAt
      properties:
        id:
          $ref: "#/components/schemas/ULID"
        eventName:
          $ref: "#/components/schemas/BehaviorEventName"
        source:
          type: string
          enum: [frontend, api, worker]
        actorType:
          type: string
          enum: [creator, receiver, system]
        creatorId:
          oneOf:
            - $ref: "#/components/schemas/ULID"
            - type: "null"
        loginId:
          type: [string, "null"]
          description: 管理员专用顶层联查字段；事件属性中不保存登录 ID。
        userSessionId:
          oneOf:
            - $ref: "#/components/schemas/ULID"
            - type: "null"
        gameId:
          oneOf:
            - $ref: "#/components/schemas/ULID"
            - type: "null"
        gameVersionId:
          oneOf:
            - $ref: "#/components/schemas/ULID"
            - type: "null"
        generationRunId:
          oneOf:
            - $ref: "#/components/schemas/ULID"
            - type: "null"
        shareId:
          oneOf:
            - $ref: "#/components/schemas/ULID"
            - type: "null"
        playSessionId:
          oneOf:
            - $ref: "#/components/schemas/ULID"
            - type: "null"
        requestId:
          oneOf:
            - $ref: "#/components/schemas/ULID"
            - type: "null"
        properties:
          $ref: "#/components/schemas/BehaviorEventProperties"
        occurredAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    BehaviorEventProperties:
      description: 具体形状由 eventName 决定，未列出的字段一律拒绝。
      anyOf:
        - $ref: "#/components/schemas/EmptyAnalyticsProperties"
        - $ref: "#/components/schemas/CreatorPageViewedProperties"
        - $ref: "#/components/schemas/TemplateAnalyticsProperties"
        - $ref: "#/components/schemas/VersionAnalyticsProperties"
        - $ref: "#/components/schemas/AssetAnalyticsProperties"
        - $ref: "#/components/schemas/GenerationSubmittedProperties"
        - $ref: "#/components/schemas/GenerationSucceededProperties"
        - $ref: "#/components/schemas/GenerationFailedProperties"
        - $ref: "#/components/schemas/ShareCreatedProperties"
        - $ref: "#/components/schemas/PublicPlayProperties"
    EmptyAnalyticsProperties:
      type: object
      additionalProperties: false
    TemplateAnalyticsProperties:
      type: object
      additionalProperties: false
      required: [templateId]
      properties:
        templateId:
          type: string
          pattern: "^[a-z][a-z0-9-]{0,63}$"
    VersionAnalyticsProperties:
      type: object
      additionalProperties: false
      required: [versionNumber, templateId]
      properties:
        versionNumber:
          type: integer
          minimum: 1
          maximum: 4294967295
        templateId:
          type: string
          pattern: "^[a-z][a-z0-9-]{0,63}$"
    AssetAnalyticsProperties:
      type: object
      additionalProperties: false
      required: [kind, mimeType, sizeBytes]
      properties:
        kind:
          type: string
          enum: [game_source, game_cover]
        mimeType:
          type: string
          minLength: 1
          maxLength: 128
        sizeBytes:
          type: integer
          minimum: 0
          maximum: 9223372036854775807
    GenerationSubmittedProperties:
      type: object
      additionalProperties: false
      required: [attemptNumber, deduplicated]
      properties:
        attemptNumber:
          type: integer
          minimum: 1
          maximum: 4294967295
        deduplicated:
          type: boolean
    GenerationSucceededProperties:
      type: object
      additionalProperties: false
      required: [executionCount]
      properties:
        executionCount:
          type: integer
          minimum: 1
          maximum: 4294967295
    GenerationFailedProperties:
      type: object
      additionalProperties: false
      required: [errorCode, retryable, executionCount]
      properties:
        errorCode:
          type: string
          enum:
            - INPUT_VALIDATION_FAILED
            - ASSET_READ_FAILED
            - GENERATION_TIMEOUT
            - PROVIDER_RATE_LIMITED
            - PROVIDER_UNAVAILABLE
            - CONTENT_REJECTED
            - ASSET_GENERATION_FAILED
            - GAME_CONFIG_ASSEMBLY_FAILED
            - GAME_CONFIG_INVALID
            - TEMPLATE_COMPATIBILITY_FAILED
            - GAME_VALIDATION_FAILED
            - STORAGE_WRITE_FAILED
            - STORAGE_CAPACITY_UNAVAILABLE
            - TASK_LEASE_EXHAUSTED
            - INTERNAL_ERROR
        retryable:
          type: boolean
        executionCount:
          type: integer
          minimum: 1
          maximum: 4294967295
    ShareCreatedProperties:
      type: object
      additionalProperties: false
      required: [lifetimeDays]
      properties:
        lifetimeDays:
          type: integer
          minimum: 1
          maximum: 90